packages/agent-sdk/pyproject.toml
[project]
name = "create-something-agents"
version = "0.1.0"
description = "CREATE SOMETHING Agent SDK - Reliable Claude agents with Anthropic's official SDK"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "CREATE SOMETHING", email = "micah@createsomething.io" }
]
keywords = ["anthropic", "claude", "agents", "ai", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "anthropic>=0.40.0",      # Official Anthropic SDK with agent support
    "httpx>=0.27.0",          # Async HTTP client for Cloudflare API
    "pyyaml>=6.0",            # YAML parsing for skills
    "pydantic>=2.0",          # Data validation
]

[project.optional-dependencies]
gemini = [
    "google-genai>=1.0.0",  # New unified Gemini SDK (replaces deprecated google-generativeai)
]
server = [
    "fastapi>=0.115.0",       # HTTP server
    "uvicorn[standard]>=0.32.0",  # ASGI server
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",            # Linting
    "mypy>=1.13.0",           # Type checking
]
all = [
    "create-something-agents[server,dev,gemini]",
]

[project.scripts]
cs-agent = "create_something_agents.cli:main"

[project.urls]
Homepage = "https://createsomething.io"
Documentation = "https://createsomething.space/learn/agent-sdk"
Repository = "https://github.com/createsomethingtoday/create-something-monorepo"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/create_something_agents"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line length handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require API keys)",
]
